% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1.R
\docType{data}
\name{govt}
\alias{govt}
\title{Indian governments by date, party, and prime minister}
\format{
A data frame with 28 rows and 7 variables:
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_prime_ministers_of_India}
}
\usage{
govt
}
\description{
A dataset containing the dates of prime minister and party of
every Indian government since the British Raj. URL may no longer
have the correct data.
}
\keyword{datasets}
