% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-clean-news-desks.R
\name{nyt_bind_news_desk_lookups}
\alias{nyt_bind_news_desk_lookups}
\alias{nyt_clean_news_desks}
\title{Clean 'news_desk' values}
\usage{
nyt_bind_news_desk_lookups(news_desk_output_path = "renamed_news_desks")

nyt_clean_news_desks(combined_df, news_desk_output_path = "renamed_news_desks")
}
\arguments{
\item{news_desk_output_path}{folder to find news desks post-cleaning}

\item{combined_df}{output of \code{nyt_clean_api_tbl()}}
}
\value{
\code{nyt_clean_news_desks()} returns a nested df where news desk values have been consolidated based on the lookup table created from the shiny app
}
\description{
Values for 'news_desk' are often slightly different spellings of the same item.
There are too many instances to fix with a \code{case_when()} statement
so use the shiny app to create a lookup table.
Then join in the replacement values. If you don't have a file of
replacements in the "renamed_news_desks" folder, it will skip that step.
}
\details{
You may not be able to write out the entire lookup table in one file
so you first need to bind together multiple files, if they exist.
}
\examples{
\dontrun{
nyt_clean_news_desks(combined_df)
}
}
