% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-clean-keywords.R
\name{nyt_bind_values_lookups}
\alias{nyt_bind_values_lookups}
\alias{nyt_clean_keywords}
\title{Clean keywords}
\usage{
nyt_bind_values_lookups(values_output_path = "renamed_values")

nyt_clean_keywords(
  unnested_df,
  values_output_path = "renamed_values",
  multi_names_input_path = "multi_named_values",
  multi_names_output_path = "single_named_values"
)
}
\arguments{
\item{values_output_path}{folder to find keyword values post-cleaning}

\item{unnested_df}{output of \code{nyt_unnest_df()}}

\item{multi_names_input_path}{folder to find keyword values with more than 1 name (category)}

\item{multi_names_output_path}{folder to save corrected keyword values csv}
}
\value{
\code{nyt_clean_keywords()} returns an unnested df with
replaced keyword values and writes out a file of all keywords
with more than one name
}
\description{
Once keywords (values) have been unnested, they need a lot of
cleaning using the same shiny app used for news desks. Once
you have joined in the replacement keywords, you also have to
ensure every keyword has one name (or category). Possible names
are "subjects", "persons", "glocations", "organizations", or
"creative_works". These cases are somewhat rare but simplify
network analysis. This file must be named "single_named_keywords.csv"
and found in the "single_named_values" folder under the top level
project folder.
}
\examples{
\dontrun{
nyt_clean_keywords(unnested_df)
}
}
