% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-build-data.R
\name{nyt_build_data}
\alias{nyt_build_data}
\title{Build data}
\usage{
nyt_build_data(
  desks_to_replace = FALSE,
  keyword_values_to_replace = FALSE,
  keyword_names_to_replace = FALSE,
  glocations_to_code = FALSE,
  ...
)
}
\arguments{
\item{desks_to_replace}{whether you need to rename desks}

\item{keyword_values_to_replace}{whether you need to rename keyword values}

\item{keyword_names_to_replace}{whether you need to rename keyword categories}

\item{glocations_to_code}{whether you need to geocode locations}

\item{...}{any other arguments}
}
\value{
The full nested_df and govt.csv required as input to the shiny app
}
\description{
Once you have finished querying data from the API, this
function will send the folder of files retrieved from the
API through a pipeline of functions to produce the output
dataset to the shiny app. You can use this if several
manual data cleaning steps (renaming desks or keyword values),
as well as having already geocoded locations, are already
complete or are being skipped.
}
\examples{
\dontrun{
nyt_build_data()
}
}
