% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-download-shiny-files.R
\name{nyt_download_shiny_files}
\alias{nyt_download_shiny_files}
\title{Download shiny app files}
\usage{
nyt_download_shiny_files(shiny_app_folder = "nyt_shiny_app")
}
\arguments{
\item{shiny_app_folder}{name of shiny app folder}
}
\value{
\code{nyt_download_shiny_files()} downloads necessary ui.R/server.R etc to run shiny app locally
}
\description{
\code{nyt_build_data()} creates the final nested dataset.
But to run the shiny app locally, you'll need the files
\href{https://github.com/seanangio/nytindia/tree/master/inst/examples/nyt_india_app}{here}
}
\examples{
\dontrun{
nyt_download_shiny_files()
}
}
