% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-fix-keywords.R
\name{nyt_fix_keywords}
\alias{nyt_fix_keywords}
\title{Fix keywords}
\usage{
nyt_fix_keywords(
  consolidated_unnested_df,
  multi_names_output_path = "single_named_values"
)
}
\arguments{
\item{consolidated_unnested_df}{output of \code{nyt_clean_keywords()}}

\item{multi_names_output_path}{folder to find corrected keyword values csv}
}
\value{
an unnested df with consolidated, non-duplicate keyword values and names
}
\description{
After choosing a name for keywords that have more than one
in a spreadsheet, join those results back in. You also need to
fix cases where the same keyword is given more than one rank.
}
\examples{
\dontrun{
nyt_fix_keywords(consolidated_unnested_df)
}
}
