% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08-add-coords-countries.R
\name{nyt_join_coords_countries}
\alias{nyt_join_coords_countries}
\title{Join in coordinates and country status for location keywords}
\usage{
nyt_join_coords_countries(
  unnested_df_values_fixed,
  mapquest_folder = "mapquest"
)
}
\arguments{
\item{unnested_df_values_fixed}{output of \code{nyt_fix_keywords()}}

\item{mapquest_folder}{where mapquest API output is found}
}
\value{
unnested df with lat-lon coords for every glocation keyword
}
\description{
Using the Mapquest API, you found lat-lon coordinates for every
glocation keyword. Join in those results to the unnested df.
Use a list of countries to identify which glocation keywords
are countries.
}
\examples{
\dontrun{
full_unnested_df <- nyt_join_coords_countries(unnested_df_values_fixed)
}
}
