% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-query-mapquest.R
\name{nyt_query_mapquest_api}
\alias{nyt_query_mapquest_api}
\title{Query Mapquest API for lat-lon coordinates}
\usage{
nyt_query_mapquest_api(
  unnested_df_values_fixed,
  mapquest_folder = "mapquest",
  api_key = Sys.getenv("MAPQUEST_KEY"),
  nyt_user_agent = Sys.getenv("NYT_USER_AGENT")
)
}
\arguments{
\item{unnested_df_values_fixed}{unnested df containing all glocation keywords}

\item{mapquest_folder}{folder where to write mapquest output}

\item{api_key}{"MAPQUEST_KEY" environmental variable}

\item{nyt_user_agent}{"NYT_USER_AGENT" environmental variable}
}
\value{
file of glocation keywords and their lat-lon coordinates
}
\description{
Many keyword values are "glocations". Here you write a file
of lat-lon coordinates for every unique glocation keyword.
}
\examples{
\dontrun{
nyt_query_mapquest_api(unnested_df_values_fixed)
}
}
