% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-re-nest-keywords.R
\name{extract_rank}
\alias{extract_rank}
\alias{find_max_kword}
\alias{nyt_re_nest_keywords}
\title{Re-nest keywords}
\usage{
extract_rank(df)

find_max_kword(df)

nyt_re_nest_keywords(full_unnested_df, max_allowable_india_rank = 30)
}
\arguments{
\item{df}{keyword df belonging to one article}

\item{full_unnested_df}{output of \code{nyt_join_coords_countries()}}

\item{max_allowable_india_rank}{upper limit of india keyword rank of an article}
}
\value{
\code{nyt_re_nest_keywords()} returns a nested df
}
\description{
Now that you have a final unnested version of the data, you need
to create a nested version. In the process you are also ready to
create new columns (india_rank, max_kword, and in_of_n_kword),
which detail the relevance of India to an article.
}
\details{
\code{extract_rank()} and \code{find_max_kword()} are small helper functions
used when iterating over each article.
}
\examples{
\dontrun{
full_nested_df <- nyt_re_nest_keywords(full_unnested_df)
}
}
