% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12-nyt_run_example.R
\name{nyt_run_example}
\alias{nyt_run_example}
\title{Run NYT India app with sample data or lookup table app}
\usage{
nyt_run_example(example)
}
\arguments{
\item{example}{either "nyt_india_app" or "lookup_table_app"}
}
\description{
Launch a Shiny app that shows a demo of what can be done
after querying and preparing data. You can also launch the
lookup table app for renaming news desks and keyword values.
}
\details{
All credit to this \href{https://github.com/daattali/shinyjs/blob/master/R/runExample.R}{shinyjs example}.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  nyt_run_example("nyt_india_app")
  nyt_run_example("lookup_table_app")
}
}
