% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-unnest-df.R
\name{nyt_unnest_df}
\alias{nyt_unnest_df}
\title{Unnest keyword values from nested df}
\usage{
nyt_unnest_df(
  nested_df,
  max_keyword_length = 200,
  values_input_path = "values_lookup_input",
  values_output_path = "renamed_values"
)
}
\arguments{
\item{nested_df}{output from \code{nyt_clean_news_desks()}}

\item{max_keyword_length}{somewhat arbitrarily set to 200 for long strings that aren't really keywords}

\item{values_input_path}{folder name to write keyword values pre-cleaning}

\item{values_output_path}{folder to find keyword values post-cleaning}
}
\value{
unnested data frame where each row is a keyword in an article instead of each row being an article
}
\description{
Up until now, each row is an article with keywords stored as a nested dataframe.
Unnesting keywords makes it easier to wrangle those values.
This function also writes out a file of keyword values to be
used as input to the lookup table shiny app and creates folder for
replacements.
}
\examples{
\dontrun{
nyt_unnest_df(nested_df)
}
}
