% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10-write-final-nested-df.R
\name{nyt_write_final_nested_df}
\alias{nyt_write_final_nested_df}
\title{Write final nested dataset for shiny app}
\usage{
nyt_write_final_nested_df(
  full_nested_df,
  shiny_app_path = "nyt_shiny_app",
  sample = FALSE,
  sample_pct = 0.1
)
}
\arguments{
\item{full_nested_df}{output of \code{nyt_re_nest_keywords()}}

\item{shiny_app_path}{name of shiny app folder}

\item{sample}{whether to write a smaller sample file from full_nested_df}

\item{sample_pct}{percentage of rows to use in the sample}
}
\value{
written nested dataframe (and sample)
}
\description{
Write final dataset, along with a sample, to the folder
that will house the app. The unnested df gets created in the app
so you don't need to create it here. It also writes the built-in
\code{govt} dataset to the same location.
}
\examples{
\dontrun{
nyt_write_final_dataset(full_nested_df)
}
}
