% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nytindia.R
\docType{package}
\name{nytindia}
\alias{nytindia}
\title{\code{nytindia} package}
\description{
Prepare and visualize India-tagged data from the NYT Article Search API
}
\details{
See the README on
\href{https://github.com/seanangio/nytindia#readme}{GitHub}
}
